feat: disable generate button during race

+++ async-race-api/src/components/winners/drivers/deleteWinner.ts
@@ -1,6 +1,11 @@
+import { getAllWinnersId } from './getAllWinnersId';
+
 export default async function deleteWinner(id: number) {
-  const url = `http://127.0.0.1:3000/winners/${id}`;
-  await fetch(url, {
-    method: 'DELETE',
-  });
+  const arrCarWinIds = await getAllWinnersId();
+  if (arrCarWinIds.includes(id)) {
+    const url = `http://127.0.0.1:3000/winners/${id}`;
+    await fetch(url, {
+      method: 'DELETE',
+    });
+  }
 }
